+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.5] - 15.12.2019
+
+### Added
+
+    - More badger options in the profiles
+    - "auto" profile chooses best setting based on available system memory
+
+### Changed
+
+    - "compactLevel0OnClose" is now disabled per default
+    - Faster shutdown of the node
+
+### Config file changes
+
+New option:
+```json
+  "useProfile": "auto",
+```
+
 ## [0.2.4] - 15.12.2019
 
 This release fixes a CRITICAL bug! You have to delete your database folder.